analysis/postprocess/requirements.txt
click==8.2.1
matplotlib==3.10.3
numpy==2.2.6
pandas==2.2.3
polars==1.30.0
pyarrow==20.0.0
PyQt5==5.15.11
rerun-sdk==0.23.2
scipy==1.15.3
